requirements.txt
requests
python-dotenv>=2.31.0
google-genai>=1.0.0
